一系列功能 诸如个人信息中心 以及管理页面查看统计 ，以及退出登录功能！（测试完毕） OK

+++ src/main/java/com/example/mfsp/entity/Address.java
@@ -18,7 +18,7 @@ public class Address {
     //null int(10)
     private Integer userid;
 
-    private Integer userphone;
+    private Long userphone;
 
     private String username;
 
@@ -43,9 +43,9 @@ public class Address {
         this.username = username;
     }
     @Column(name ="userphone")
-    public Integer getUserphone(){ return userphone;}
+    public Long getUserphone(){ return userphone;}
 
-    public void setUserphone(Integer userphone) {
+    public void setUserphone(Long userphone) {
         this.userphone = userphone;
     }
